src/third_party/WebKit/LayoutTests/external/wpt/custom-elements/range-and-constructors-expected.txt

This is a testharness.js-based test.
FAIL Range.cloneContents should invoke constructor in tree order assert_array_equals: property 0, expected "root-0" but got "root-0-0"
FAIL Range.extractContents should invoke constructor in tree order assert_array_equals: property 0, expected "root-0" but got "root-0-0"
Harness: the test ran to completion.
